chatbridge renamed accurately to webbridge renamed files to match framework convention (less underscores)
