+++ hangupsbot/threadmanager.py
